# Optimize `Object#instance_variable_names`