--- a/activesupport/lib/active_support/core_ext/object/instance_variables.rb
+++ b/activesupport/lib/active_support/core_ext/object/instance_variables.rb
@@ -15,16 +15,24 @@ class Object
     Hash[instance_variables.map { |name| [name[1..-1], instance_variable_get(name)] }]
   end
 
-  # Returns an array of instance variable names as strings including "@".
-  #
-  #   class C
-  #     def initialize(x, y)
-  #       @x, @y = x, y
-  #     end
-  #   end
-  #
-  #   C.new(0, 1).instance_variable_names # => ["@y", "@x"]
-  def instance_variable_names
-    instance_variables.map(&:to_s)
+  if Symbol.method_defined?(:name) # RUBY_VERSION >= "3.0"
+    # Returns an array of instance variable names as strings including "@".
+    #
+    #   class C
+    #     def initialize(x, y)
+    #       @x, @y = x, y
+    #     end
+    #   end
+    #
+    #   C.new(0, 1).instance_variable_names # => ["@y", "@x"]
+    def instance_variable_names
+      instance_variables.map(&:name)
+    end
+  else
+    def instance_variable_names
+      variables = instance_variables
+      variables.map! { |s| s.to_s.freeze }
+      variables
+    end
   end
 end